+++ src/bench/rpc_mempool.cpp
@@ -19,7 +19,7 @@ static void AddTx(const CTransactionRef &tx, const Amount &fee,
     pool.addUnchecked(CTxMemPoolEntry(tx, fee, /* time */ 0,
                                       /* height */ 1,
                                       /* spendsCoinbase */ false,
-                                      /* sigOpCount */ 1, lp));
+                                      /* sigChecks */ 1, lp));
 }
 
 static void RPCMempoolVerbose(benchmark::State &state) {