+++ src/validation.cpp
@@ -3930,10 +3930,6 @@ static bool ContextualCheckBlock(const CBlock &block, CValidationState &state,
     // Check transactions:
     // - canonical ordering
     // - ensure they are finalized
-    // - perform a preliminary block-sigops count (they will be recounted more
-    // strictly during ConnectBlock).
-    // - perform a transaction-sigops check (again, a more strict check will
-    // happen in ConnectBlock).
     const CTransaction *prevTx = nullptr;
     for (const auto &ptx : block.vtx) {
         const CTransaction &tx = *ptx;